refactor: renamed some dto files
