+++ src/modules/api-key/controllers/api-key.admin.controller.ts
@@ -46,10 +46,10 @@ import {
 import { ApiKeyCreateRequestDto } from '@modules/api-key/dtos/request/api-key.create.request.dto';
 import { ApiKeyUpdateDateRequestDto } from '@modules/api-key/dtos/request/api-key.update-date.request.dto';
 import { ApiKeyUpdateRequestDto } from '@modules/api-key/dtos/request/api-key.update.request.dto';
-import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.dto';
+import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.response.dto';
 import { ApiKeyGetResponseDto } from '@modules/api-key/dtos/response/api-key.get.response.dto';
 import { ApiKeyListResponseDto } from '@modules/api-key/dtos/response/api-key.list.response.dto';
-import { ApiKeyResetResponseDto } from '@modules/api-key/dtos/response/api-key.reset.dto';
+import { ApiKeyResetResponseDto } from '@modules/api-key/dtos/response/api-key.reset.response.dto';
 import { ApiKeyParsePipe } from '@modules/api-key/pipes/api-key.parse.pipe';
 import { ApiKeyDoc } from '@modules/api-key/repository/entities/api-key.entity';
 import { ApiKeyService } from '@modules/api-key/services/api-key.service';
