--- src/modules/api-key/dtos/response/api-key.reset.dto.ts
@@ -1,3 +1,0 @@
-import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.dto';
-
-export class ApiKeyResetResponseDto extends ApiKeyCreateResponseDto {}
