+++ src/modules/api-key/dtos/response/api-key.reset.response.dto.ts
@@ -1,0 +1,3 @@
+import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.response.dto';
+
+export class ApiKeyResetResponseDto extends ApiKeyCreateResponseDto {}
