+++ src/modules/api-key/services/api-key.service.ts
@@ -19,10 +19,10 @@ import {
 } from '@modules/api-key/dtos/request/api-key.create.request.dto';
 import { ApiKeyUpdateDateRequestDto } from '@modules/api-key/dtos/request/api-key.update-date.request.dto';
 import { ApiKeyUpdateRequestDto } from '@modules/api-key/dtos/request/api-key.update.request.dto';
-import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.dto';
+import { ApiKeyCreateResponseDto } from '@modules/api-key/dtos/response/api-key.create.response.dto';
 import { ApiKeyGetResponseDto } from '@modules/api-key/dtos/response/api-key.get.response.dto';
 import { ApiKeyListResponseDto } from '@modules/api-key/dtos/response/api-key.list.response.dto';
-import { ApiKeyResetResponseDto } from '@modules/api-key/dtos/response/api-key.reset.dto';
+import { ApiKeyResetResponseDto } from '@modules/api-key/dtos/response/api-key.reset.response.dto';
 import { IApiKeyService } from '@modules/api-key/interfaces/api-key.service.interface';
 import {
     ApiKeyDoc,
