+++ src/modules/user/dtos/request/user.check.request.dto.ts
@@ -1,6 +1,6 @@
 import { PickType } from '@nestjs/swagger';
 import { UserCreateRequestDto } from '@modules/user/dtos/request/user.create.request.dto';
-import { UserUpdateClaimUsernameRequestDto } from '@modules/user/dtos/request/user.update-claim-username.dto';
+import { UserUpdateClaimUsernameRequestDto } from '@modules/user/dtos/request/user.update-claim-username.request.dto';
 
 export class UserCheckUsernameRequestDto extends UserUpdateClaimUsernameRequestDto {}
